refactor: split PDF view controller responsibilities
