+++ YetAnotherEBookReader/Views/PDFView/YabrPDFViewController.swift
@@ -47,9 +47,17 @@ class YabrPDFViewController: UIViewController, UIGestureRecognizerDelegate, Obse
     weak var readerEngineDelegate: ReaderEngineDelegate?
     var initialPosition: ReaderEnginePosition?
     
-    var annotationManager: PDFAnnotationManager!
-    var bookmarkManager: PDFBookmarkManager!
-    var searchController: PDFSearchController!
+    lazy var annotationManager: PDFAnnotationManager = {
+        let bookId = yabrPDFMetaSource?.yabrPDFBook(pdfView, info: "Key") ?? ""
+        return PDFAnnotationManager(pdfView: pdfView, delegate: readerEngineDelegate, bookId: bookId)
+    }()
+    lazy var bookmarkManager = PDFBookmarkManager(pdfView: pdfView, metaSource: yabrPDFMetaSource)
+    lazy var searchController = PDFSearchController(pdfView: pdfView, metaSource: yabrPDFMetaSource)
+    lazy var marginCropController = PDFMarginCropController(
+        pdfView: pdfView,
+        blankView: blankView,
+        blankActivityView: blankActivityView
+    )
     
     @Published var pdfOptions = PDFOptions() {
         didSet {
@@ -115,7 +123,7 @@ class YabrPDFViewController: UIViewController, UIGestureRecognizerDelegate, Obse
                 }
                 
                 if pdfOptions.pageMode == .Page {
-                    self.pageVisibleContentBounds.removeAll()
+                    self.marginCropController.clearCache()
                 }
             }
             
@@ -130,8 +138,6 @@ class YabrPDFViewController: UIViewController, UIGestureRecognizerDelegate, Obse
     }
         
     var pageViewPositionHistory = [Int: PageViewPosition]() //key is 1-based (pdfView.currentPage?.pageRef?.pageNumber)
-    
-    var pageVisibleContentBounds: [PageVisibleContentKey: PageVisibleContentValue] = [:]
     
     func open() -> Int {
         guard let pdfURL = yabrPDFMetaSource?.yabrPDFURL(pdfView) else { return -1 }
@@ -170,11 +176,6 @@ class YabrPDFViewController: UIViewController, UIGestureRecognizerDelegate, Obse
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        let bookId = self.yabrPDFMetaSource?.yabrPDFBook(self.pdfView, info: "Key") ?? ""
-        self.annotationManager = PDFAnnotationManager(pdfView: self.pdfView, delegate: self.readerEngineDelegate, bookId: bookId)
-        self.bookmarkManager = PDFBookmarkManager(pdfView: self.pdfView, metaSource: self.yabrPDFMetaSource)
-        self.searchController = PDFSearchController(pdfView: self.pdfView, metaSource: self.yabrPDFMetaSource)
-        
         // pdfView.usePageViewController(true, withViewOptions: nil)
         
         NotificationCenter.default.addObserver(self, selector: #selector(handlePageChange(notification:)), name: .PDFViewPageChanged, object: pdfView)
@@ -183,375 +184,17 @@ class YabrPDFViewController: UIViewController, UIGestureRecognizerDelegate, Obse
         
         pdfView.autoScales = false
 
-        let backgroundColor = UIColor(cgColor: pdfOptions.fillColor)
-        self.navigationController?.navigationBar.barTintColor = backgroundColor
-        self.navigationController?.navigationBar.backgroundColor = backgroundColor
-        self.navigationController?.toolbar.barTintColor = backgroundColor
-        self.navigationController?.toolbar.backgroundColor = backgroundColor
-
-        self.tabBarController?.tabBar.barTintColor = backgroundColor
-        self.tabBarController?.tabBar.backgroundColor = backgroundColor
-        
-        pageIndicator.setTitle("0 / 0", for: .normal)
-        pageIndicator.addAction(UIAction(handler: { [self] (action) in
-            guard let curPageNum = pdfView.currentPage?.pageRef?.pageNumber,
-                  let bounds = pageVisibleContentBounds[
-                    PageVisibleContentKey(
-                        pageNumber: curPageNum,
-                        readingDirection: pdfOptions.readingDirection,
-                        hMarginDetectStrength: pdfOptions.hMarginDetectStrength,
-                        vMarginDetectStrength: pdfOptions.vMarginDetectStrength
-                    )
-                  ],
-                  let image = bounds.thumbImage else { return }
-            self.thumbImageView.image = image
-            
-//            self.present(thumbController, animated: true, completion: nil)
-            
-            let pageController = YabrPDFNavigationPageVC()
-            pageController.pdfViewController = self
-            pageController.yabrPDFView = self.pdfView
-            pageController.yabrPDFMetaSource = self.yabrPDFMetaSource
-            
-            let nav = UINavigationController(rootViewController: pageController)
-            if let fillColor = PDFPageWithBackground.fillColor {
-                nav.navigationBar.backgroundColor = UIColor(cgColor: fillColor)
-                nav.navigationBar.barTintColor = UIColor(cgColor: fillColor)
-            }
-            
-            self.present(nav, animated: true)
-            
-        }), for: .primaryActionTriggered)
-        
-        pageSlider.minimumValue = 1
-        pageSlider.maximumValue = Float(pdfView.document?.pageCount ?? 1)
-        pageSlider.isContinuous = true
-        pageSlider.addAction(UIAction(handler: { (action) in
-            guard let currentPageNumber = self.pdfView.currentPage?.pageRef?.pageNumber else { return }
-            let destPageNumber = Int(self.pageSlider.value.rounded())
-            print("\(#function) current=\(currentPageNumber) target=\(destPageNumber)")
-            
-            guard currentPageNumber != destPageNumber,
-                  let destPage = self.pdfView.document?.page(at: destPageNumber - 1) else { return }
-            
-            self.addBlankSubView(page: destPage)
-            self.pdfView.go(to: destPage)
-        }), for: .valueChanged)
-
-        pagePrevButton.setImage(UIImage(systemName: "arrow.left"), for: .normal)
-        pagePrevButton.addAction(UIAction(handler: { (action) in
-            self.updatePageViewPositionHistory()
-            self.addBlankSubView(page: self.pdfView.currentPage)    //FIXME get actual page
-
-            if self.pdfView.displaysRTL {
-                self.pdfView.goToNextPage(self.pagePrevButton)
-            } else {
-                self.pdfView.goToPreviousPage(self.pagePrevButton)
-            }
-        }), for: .primaryActionTriggered)
-        
-        pageNextButton.setImage(UIImage(systemName: "arrow.right"), for: .normal)
-        pageNextButton.addAction(UIAction(handler: { (action) in
-            self.updatePageViewPositionHistory()
-            self.addBlankSubView(page: self.pdfView.currentPage)    //FIXME get actual page
-
-            if self.pdfView.displaysRTL {
-                self.pdfView.goToPreviousPage(self.pagePrevButton)
-            } else {
-                self.pdfView.goToNextPage(self.pagePrevButton)
-            }
-        }), for: .primaryActionTriggered)
-        
-        pageBackButton.setImage(UIImage(systemName: "arrow.uturn.left"), for: .normal)
-        let tintColor: UIColor = pdfOptions.isDark(.lightText, .darkText)
-        pageBackButton.setTitleColor(tintColor, for: .normal)
-        
-        stackView.distribution = .fill
-        stackView.alignment = .fill
-        stackView.axis = .horizontal
-        stackView.spacing = 16.0
-        
-        if #available(iOS 16.0, *) {
-            pageBackButton.isHidden = true
-            stackView.addArrangedSubview(pageBackButton)
-            
-            let pageBackAction = UIAction(handler: { action in
-                guard let historyItem = self.historyMenu.children.last as? UIAction
-                else {
-                    return
-                }
-                self.addBlankSubView(page: self.pdfView.currentPage)    //FIXME get actual page
-                historyItem.performWithSender(self, target: self.pdfView)
-            })
-            
-            pageBackButton.addAction(pageBackAction, for: .primaryActionTriggered)
-        }
-        
-        pageAuxButton.setImage(UIImage(systemName: "square.split.bottomrightquarter"), for: .normal)
-        pageAuxButton.addAction(.init(handler: { [self] action in
-            if pdfViewAux.superview == nil {
-                pdfViewAux.backgroundColor = pdfView.backgroundColor
-//                pdfViewAux.layer.shadowColor = pdfView.backgroundColor.cgColor
-                
-                pdfViewAux.frame = .init(
-                    origin: .init(x: 150.0, y: view.frame.height - 260),
-                    size: .init(width: pdfView.frame.width - 200.0, height: 200.0)
-                )
-                
-                if pdfViewAux.document == nil {
-                    pdfViewAux.document = pdfView.document
-                    
-                    pdfViewAux.layer.borderWidth = 2
-                    pdfViewAux.layer.cornerRadius = 8
-                    pdfViewAux.layer.shadowRadius = 16
-                    
-                    pdfViewAux.scaleFactor = pdfView.scaleFactor * 0.8
-                    pdfViewAux.displayMode = .singlePageContinuous
-                    pdfViewAux.displayDirection = .vertical
-                    pdfViewAux.interpolationQuality = .high
-                    
-                    if let currentDestination = pdfView.currentDestination {
-                        pdfViewAux.go(to: currentDestination)
-                    }
-                }
-                
-                view.addSubview(pdfViewAux)
-            } else {
-                pdfViewAux.removeFromSuperview()
-            }
-            
-        }), for: .primaryActionTriggered)
-        
-        stackView.addArrangedSubview(pagePrevButton)
-        stackView.addArrangedSubview(pageSlider)
-        stackView.addArrangedSubview(pageIndicator)
-        stackView.addArrangedSubview(pageNextButton)
-        stackView.addArrangedSubview(pageAuxButton)
-        
-        let toolbarView = UIBarButtonItem(customView: stackView)
-        setToolbarItems([toolbarView], animated: false)
-        
-        print("stackView \(self.navigationController?.view.frame ?? .zero) \(self.navigationController?.toolbar.frame ?? .zero)")
-        stackView.frame = self.navigationController?.toolbar.frame ?? .zero
-        
-        navigationItem.setLeftBarButtonItems([
-            UIBarButtonItem(image: UIImage(systemName: "xmark.circle"), style: .done, target: self, action: #selector(finishReading(sender:))),
-            UIBarButtonItem(title: "Navigations", image: UIImage(systemName: "list.bullet"), primaryAction: UIAction(handler: { action in
-                let navigationController = YabrPDFNavigationPageVC()
-                navigationController.pdfViewController = self
-                navigationController.yabrPDFView = self.pdfView
-                navigationController.yabrPDFMetaSource = self.yabrPDFMetaSource
-                
-                let nav = UINavigationController(rootViewController: navigationController)
-                if let fillColor = PDFPageWithBackground.fillColor {
-                    nav.navigationBar.backgroundColor = UIColor(cgColor: fillColor)
-                    nav.navigationBar.barTintColor = UIColor(cgColor: fillColor)
-                }
-                
-                self.present(nav, animated: true)
-            })),
-            UIBarButtonItem(title: "Annotations", image: UIImage(systemName: "bookmark"), primaryAction: UIAction(handler: { action in
-                let annotationController = YabrPDFAnnotationPageVC()
-                annotationController.pdfViewController = self
-                annotationController.yabrPDFView = self.pdfView
-                annotationController.yabrPDFMetaSource = self.yabrPDFMetaSource
-                
-                let nav = UINavigationController(rootViewController: annotationController)
-                if let fillColor = PDFPageWithBackground.fillColor {
-                    nav.navigationBar.backgroundColor = UIColor(cgColor: fillColor)
-                    nav.navigationBar.barTintColor = UIColor(cgColor: fillColor)
-                }
-                
-                self.present(nav, animated: true)
-            }))
-//            UIBarButtonItem(title: "Zoom Out", image: UIImage(systemName: "minus.magnifyingglass"), primaryAction: UIAction(handler: { (UIAction) in
-//                self.pdfView.scaleFactor = (self.pdfView.scaleFactor ) / 1.1
-//            })),
-//            UIBarButtonItem(title: "Zoom In", image: UIImage(systemName: "plus.magnifyingglass"), primaryAction: UIAction(handler: { (UIAction) in
-//                self.pdfView.scaleFactor = (self.pdfView.scaleFactor ) * 1.1
-//            })),
-        ], animated: true)
-        
-        let shareOriginalPDF = UIAction(title: "Original PDF") { [self] action in
-            print("\(#function) \(action)")
-            sharePDF(annotated: false)
-        }
-        
-        let shareAnnotatedPDF = UIAction(title: "Annotated PDF") { [self] action in
-            sharePDF(annotated: true)
-        }
-        
-        shareBarButtonItem.title = "Share"
-        shareBarButtonItem.image = UIImage(systemName: "square.and.arrow.up")
-        shareBarButtonItem.menu = UIMenu(children: [shareOriginalPDF, shareAnnotatedPDF])
-        
-        navigationItem.setRightBarButtonItems([
-            UIBarButtonItem(image: UIImage(systemName: "clock"), menu: historyMenu),
-            UIBarButtonItem(
-                title: "Options",
-                image: UIImage(systemName: "doc.badge.gearshape"),
-                primaryAction: UIAction { (UIAction) in
-                    //self.present(self.optionMenu, animated: true, completion: nil)
-                    let optionView = PDFOptionView(
-                        pdfViewController: Binding<YabrPDFViewController>(
-                            get: { return self },
-                            set: { _ in }
-                        ),
-                        pdfOptions: self.pdfOptions
-                    )
-                    
-                    let optionViewController = UIHostingController(rootView: optionView.fixedSize())
-                    optionViewController.preferredContentSize = CGSize(width:340, height:700)
-                    optionViewController.modalPresentationStyle = .popover
-                    optionViewController.popoverPresentationController?.barButtonItem = self.navigationItem.rightBarButtonItems?[1]
-                    
-                    self.present(optionViewController, animated: true, completion: nil)
-                }
-            ),
-            shareBarButtonItem
-        ], animated: true)
-        self.navigationItem.rightBarButtonItems?.first?.isEnabled = false
-        
-        buildTocList()
-        
-        let docTitle = self.pdfView.document?.documentAttributes?[PDFDocumentAttribute.titleAttribute] as? String
-        titleInfoButton.setTitle(docTitle ?? "", for: .normal)
-        titleInfoButton.contentHorizontalAlignment = .center
-        titleInfoButton.showsMenuAsPrimaryAction = true
-        titleInfoButton.frame = CGRect(x:0, y:0, width: navigationController?.navigationBar.frame.width ?? 600 / 2, height: 40)
-        
-        navigationItem.titleView = titleInfoButton
-        
-        pdfView.delegate = self
-        
-        pdfView.translatesAutoresizingMaskIntoConstraints = false
-        
-        self.view.addSubview(pdfView)
-        
-        NSLayoutConstraint.activate([
-            pdfView.topAnchor.constraint(equalTo: self.view.topAnchor),
-            pdfView.bottomAnchor.constraint(
-                equalTo: self.view.bottomAnchor,
-                constant: -self.stackView.frame.height
-            ),
-            pdfView.leftAnchor.constraint(equalTo: self.view.leftAnchor),
-            pdfView.rightAnchor.constraint(equalTo: self.view.rightAnchor)
-        ])
-
-        let defaultMenuItems = buildDefaultMenuItems()
-        UIMenuController.shared.menuItems = defaultMenuItems
-        UIMenuController.shared.update()
-        
-        UIMenuController.installTo(responder: self.pdfView)
-        
-        if #available(iOS 16.0, *) {
-//            pdfView.interactions.forEach {
-//                pdfView.removeInteraction($0)
-//            }
-//
-//            pdfView.gestureRecognizers?.forEach {
-//                pdfView.removeGestureRecognizer($0)
-//            }
-//
-//            print("\(#function) interactions=\(pdfView.interactions.count)")
-            
-//                let newEditMenu = UIEditMenuInteraction(delegate: self)
-//                pdfView.addInteraction(newEditMenu)
-//                let longPress = UILongPressGestureRecognizer(target: self, action: #selector(didLongPress(_:)))
-//
-//                longPress.allowedTouchTypes = [UITouch.TouchType.direct.rawValue as NSNumber]
-//                pdfView.addGestureRecognizer(longPress)
-            
-            
-//            let testLabel = UILabel()
-//            testLabel.text = "Highlight"
-//            testLabel.textColor = .black
-//            testLabel.backgroundColor = .gray
-//            testLabel.frame = .init(x: 100, y: 100, width: 100, height: 100)
-//            testLabel.isHidden = true
-//            self.view.addSubview(testLabel)
-            
-            self.annotationView.isHidden = true
-            self.view.addSubview(self.annotationView)
-            
-//            self.annotationView.highlightButton.addAction(.init(handler: { [self] action in
-//                highlightAction(pdfView)
-//            }), for: .primaryActionTriggered)
-            
-            self.annotationView.underlineButton.addTarget(nil, action: #selector(highlightAction(_:)), for: .primaryActionTriggered)
-            self.annotationView.highlightButton.addTarget(nil, action: #selector(highlightAction(_:)), for: .primaryActionTriggered)
-            self.annotationView.dictViewerButton.addTarget(nil, action: #selector(dictViewerAction(_:)), for: .primaryActionTriggered)
-            
-            NotificationCenter.default.addObserver(forName: .PDFViewSelectionChanged, object: pdfView, queue: nil) { [self] notification in
-//                newEditMenu.presentEditMenu(with: .init(identifier: nil, sourcePoint: .zero))
-                
-                guard let selection = pdfView.currentSelection,
-                      let selectionString = selection.string,
-                      selectionString.count > 0
-                else {
-                    annotationView.isHidden = true
-                    return
-                }
-                
-                guard let selectionLastLine = selection.selectionsByLine().last,
-                      let selectionLastLinePage = selectionLastLine.pages.last
-                else {
-                    annotationView.isHidden = true
-                    return
-                }
-                
-                let selectionBound = selectionLastLine.bounds(for: selectionLastLinePage)
-                let selectionInView = pdfView.convert(selectionBound, from: selectionLastLinePage)
-                
-                let buttonSize = CGFloat(48)
-                let padding = CGFloat(32)
-                
-                let annotationViewSize = CGSize(width: buttonSize, height: CGFloat(annotationView.arrangedSubviews.count) * buttonSize)
-                
-                var annotationViewPosition = CGPoint(
-                    x: selectionInView.maxX + padding / 2.0,
-                    y: selectionInView.maxY + padding / 2.0
-                )
-                
-                if annotationViewPosition.x + annotationViewSize.width + padding > pdfView.frame.width {
-                    annotationViewPosition.x = pdfView.frame.width - buttonSize - padding
-                }
-                
-                if annotationViewPosition.y + annotationViewSize.height + padding > pdfView.frame.height {
-                    annotationViewPosition.y = selectionInView.minY - CGFloat(annotationView.arrangedSubviews.count) * buttonSize - padding * 2.0
-                }
-                
-                annotationView.frame = .init(origin: annotationViewPosition, size: annotationViewSize)
-                
-                annotationView.backgroundColor = pdfOptions.isDark(
-                    UIColor.black.withAlphaComponent(0.9),
-                    UIColor.white.withAlphaComponent(0.9)
-                )
-                
-                annotationView.isHidden = false
-            }
-            
-        }
+        configureReaderChrome()
+        configureSelectionOverlay()
         
         self.annotationManager.injectAllHighlights()
         
         
-        blankView.contentMode = .scaleAspectFill
-        blankView.addSubview(blankActivityView)
-        pdfView.addSubview(blankView)
-        
-        blankView.layer.compositingFilter = "darkenBlendMode"
+        marginCropController.configureBlankOverlay()
         
         pdfView.prepareActions(pageNextButton: pageNextButton, pagePrevButton: pagePrevButton)
         
-        self.thumbImageView.translatesAutoresizingMaskIntoConstraints = false
-        self.thumbController.view.addSubview(self.thumbImageView)
-        NSLayoutConstraint.activate([
-            self.thumbImageView.topAnchor.constraint(equalTo: self.thumbController.view.topAnchor),
-            self.thumbImageView.bottomAnchor.constraint(equalTo: self.thumbController.view.bottomAnchor),
-            self.thumbImageView.leadingAnchor.constraint(equalTo: self.thumbController.view.leadingAnchor),
-            self.thumbImageView.trailingAnchor.constraint(equalTo: self.thumbController.view.trailingAnchor)
-        ])
+        configureThumbnailPreview()
     }
     
     override func viewDidAppear(_ animated: Bool) {
@@ -603,1009 +246,12 @@ class YabrPDFViewController: UIViewController, UIGestureRecognizerDelegate, Obse
         }
     }
     
-    func buildTocList() {
-        DispatchQueue.global(qos: .utility).async {
-            var tableOfContents = [UIMenuElement]()
-            
-            if let pdfDoc = self.pdfView.document, var outlineRoot = pdfDoc.outlineRoot {
-                while outlineRoot.numberOfChildren == 1 {
-                    outlineRoot = outlineRoot.child(at: 0)!
-                }
-                for i in (0..<outlineRoot.numberOfChildren) {
-                    self.tocList.append((outlineRoot.child(at: i)?.label ?? "Label at \(i)", outlineRoot.child(at: i)?.destination?.page?.pageRef?.pageNumber ?? 1))
-                    tableOfContents.append(UIAction(title: outlineRoot.child(at: i)?.label ?? "Label at \(i)") { (action) in
-                        guard let dest = outlineRoot.child(at: i)?.destination,
-                              let curPage = self.pdfView.currentPage
-                        else { return }
-                        
-                        self.updateHistoryMenu(curPage: curPage)
-                        
-                        if dest.page?.pageRef?.pageNumber != self.pdfView.currentPage?.pageRef?.pageNumber {
-                            self.addBlankSubView(page: dest.page)
-                        }
-                        self.pdfView.go(to: dest)
-                    })
-                    
-                }
-            }
-            
-            let navContentsMenu = UIMenu(title: "Contents", children: tableOfContents)
-            
-            DispatchQueue.main.async {
-                self.titleInfoButton.menu = navContentsMenu
-            }
-        }
-    }
-    
-    func updateHistoryMenu(curPage: PDFPage, location: CGRect? = nil) {
-        guard let pdfDoc = curPage.document else { return }
-        
-        var lastHistoryLabel = "Page \(curPage.pageRef!.pageNumber)"
-        if let outlineRoot = pdfDoc.outlineRoot,
-           let curPageSelection = curPage.selection(for: curPage.bounds(for: .mediaBox)),
-           let curPageSelectionText = curPageSelection.string,
-           curPageSelectionText.count > 5,
-           var curPageOutlineItem = pdfDoc.outlineItem(for: curPageSelection) {
-            
-            print("\(curPageSelectionText)")
-            while( curPageOutlineItem.parent != nil && curPageOutlineItem.parent != outlineRoot) {
-                curPageOutlineItem = curPageOutlineItem.parent!
-            }
-            lastHistoryLabel += " of \(curPageOutlineItem.label!)"
-        }
-        
-        var historyItems = self.historyMenu.children
-        historyItems.append(UIAction(title: lastHistoryLabel) { (action) in
-            var children = self.historyMenu.children
-            if let index = children.firstIndex(of: action) {
-                children.removeLast(children.count - index)
-                self.historyMenu = self.historyMenu.replacingChildren(children)
-                let newMenu = self.navigationItem.rightBarButtonItems?.first?.menu?.replacingChildren(children)
-                self.navigationItem.rightBarButtonItems?.first?.menu = nil  //MUST HAVE, otherwise no effect
-                self.navigationItem.rightBarButtonItems?.first?.menu = newMenu
-                if children.isEmpty {
-                    self.navigationItem.rightBarButtonItems?.first?.isEnabled = false
-                    self.pageBackButton.isHidden = true
-                    self.pageBackButton.setTitle("", for: .normal)
-                } else if let lastHistoryItem = children.last as? UIAction {
-                    self.pageBackButton.isHidden = false
-                    if lastHistoryItem.title.count > 20 {
-                        self.pageBackButton.setAttributedTitle(
-                            .init(
-                                string: " Back to \(lastHistoryItem.title.prefix(20))...",
-                                attributes: [
-                                    NSAttributedString.Key.font: UIFont.systemFont(ofSize: 12.0)
-                                ]
-                            ),
-                            for: .normal
-                        )
-                    } else {
-                        self.pageBackButton.setAttributedTitle(
-                            .init(
-                                string: " Back to \(lastHistoryItem.title)",
-                                attributes: [
-                                    NSAttributedString.Key.font: UIFont.systemFont(ofSize: 12.0)
-                                ]
-                            ),
-                            for: .normal
-                        )
-                    }
-                }
-            }
-            if curPage.pageRef?.pageNumber != self.pdfView.currentPage?.pageRef?.pageNumber {
-                self.addBlankSubView(page: curPage)
-            }
-            if let location = location {
-                self.pdfView.go(to: location, on: curPage)
-            } else {
-                self.pdfView.go(to: curPage)
-            }
-        })
-        
-        self.historyMenu = self.historyMenu.replacingChildren(historyItems)
-        let newMenu = self.navigationItem.rightBarButtonItems?.first?.menu?.replacingChildren(historyItems)
-        self.navigationItem.rightBarButtonItems?.first?.menu = nil  //MUST HAVE, otherwise no effect
-        self.navigationItem.rightBarButtonItems?.first?.menu = newMenu
-        self.navigationItem.rightBarButtonItems?.first?.isEnabled = true
-        
-        self.pageBackButton.isHidden = false
-        if lastHistoryLabel.count > 20 {
-            self.pageBackButton.setAttributedTitle(
-                .init(
-                    string: " Back to \(lastHistoryLabel.prefix(20))...",
-                    attributes: [
-                        NSAttributedString.Key.font: UIFont.systemFont(ofSize: 12.0)
-                    ]
-                    ),
-                for: .normal
-            )
-        } else {
-            self.pageBackButton.setAttributedTitle(
-                .init(
-                    string: " Back to \(lastHistoryLabel)",
-                    attributes: [
-                        NSAttributedString.Key.font: UIFont.systemFont(ofSize: 12.0)
-                    ]
-                ),
-                for: .normal
-            )
-        }
-    }
-    
-    @objc private func handlePageChange(notification: Notification) {
-        var titleLabel = initialPosition?.chapterName
-        guard let curPage = pdfView.currentPage else { return }
-
-        if var outlineRoot = pdfView.document?.outlineRoot {
-            while outlineRoot.numberOfChildren == 1 {
-                outlineRoot = outlineRoot.child(at: 0)!
-            }
-            if let curPageSelection = curPage.selection(for: curPage.bounds(for: .mediaBox)),
-               !curPageSelection.selectionsByLine().isEmpty,
-               var curPageOutlineItem = pdfView.document?.outlineItem(for: curPageSelection) {
-                while( curPageOutlineItem.parent != nil && curPageOutlineItem.parent != outlineRoot) {
-                    curPageOutlineItem = curPageOutlineItem.parent!
-                }
-                if curPageOutlineItem.label != nil && !curPageOutlineItem.label!.isEmpty {
-                    titleLabel = curPageOutlineItem.label
-                }
-            }
-        }
-        self.titleInfoButton.setTitle(titleLabel, for: .normal)
-        
-        let curPageNum = pdfView.currentPage?.pageRef?.pageNumber ?? 1
-        pageIndicator.setTitle("\(curPageNum) / \(pdfView.document?.pageCount ?? 1)", for: .normal)
-        pageSlider.setValue(Float(curPageNum), animated: true)
-        
-        print("\(#function) curPageNum=\(curPageNum) pageIndicator=\(pageIndicator.title(for: .normal) ?? "Untitled") pageSlider=\(pageSlider.value)")
-        
-        guard pdfView.frame.width > 1.0 else { return }     // have not been populated, cannot fit content
-        
-        if pdfView.displayMode != .singlePage {
-            pdfView.scaleFactor = pdfOptions.lastScale
-            
-            if let pageViewPosition = getPageViewPositionHistory(curPageNum),
-               pageViewPosition.scaler > 0,
-               pageViewPosition.viewSize == pdfView.frame.size || pageViewPosition.viewSize == .zero {
-                let lastDest = PDFDestination(
-                    page: curPage,
-                    at: pageViewPosition.point
-                )
-                lastDest.zoom = pageViewPosition.scaler
-                print("\(#function) displayMode=\(pdfView.displayMode) BEFORE POINT lastDestPoint=\(lastDest.point)")
-                
-                pdfView.scaleFactor = pageViewPosition.scaler
-                
-                pageViewPositionHistory.removeValue(forKey: curPageNum)     //prevent further application
-                pdfView.go(to: lastDest)
-            }
-            
-            return
-        }
-        
-        addBlankSubView(page: curPage)
-        
-        [PDFDisplayBox.mediaBox, PDFDisplayBox.cropBox, PDFDisplayBox.trimBox, PDFDisplayBox.bleedBox, PDFDisplayBox.artBox].forEach {
-            let bounds = curPage.bounds(for: $0)
-            print("\(#function) boundsForBox box=\($0.rawValue) bounds=\(bounds)")
-        }
-        let boundsForCropBox = curPage.bounds(for: .cropBox)
-        let boundsForMediaBox = curPage.bounds(for: .mediaBox)
-        let boundForVisibleContentKey = PageVisibleContentKey(
-            pageNumber: curPageNum,
-            readingDirection: pdfOptions.readingDirection,
-            hMarginDetectStrength: pdfOptions.hMarginDetectStrength,
-            vMarginDetectStrength: pdfOptions.vMarginDetectStrength
-        )
-        
-        let boundForVisibleContent = pageVisibleContentBounds[boundForVisibleContentKey]?.bounds ?? { () -> CGRect in
-            let bounds = getVisibleContentsBound(pdfPage: curPage, readingDirection: pdfOptions.readingDirection, hMarginDetectStrength: pdfOptions.hMarginDetectStrength, vMarginDetectStrength: pdfOptions.vMarginDetectStrength, marginOffset: pdfOptions.marginOffset)
-            pageVisibleContentBounds[boundForVisibleContentKey] = bounds
-            return bounds.bounds
-        }()
-        pageVisibleContentBounds[boundForVisibleContentKey]?.lastUsed = Date()
-        
-        //pre-analyze next page
-        while( pageVisibleContentBounds.count > 9 ) {
-            if let minPageEntry = pageVisibleContentBounds.min(by: {$0.value.lastUsed < $1.value.lastUsed}) {
-                pageVisibleContentBounds.removeValue(forKey: minPageEntry.key)
-                print("\(#function) pageVisibleContentBounds.removeValue=\(minPageEntry.key.pageNumber)")
-            } else {
-                break
-            }
-        }
-        let readingDirection = pdfOptions.readingDirection
-        let hMarginDetectStrength = pdfOptions.hMarginDetectStrength
-        let vMarginDetectStrength = pdfOptions.vMarginDetectStrength
-        let marginOffset = pdfOptions.marginOffset
-        
-        DispatchQueue.global(qos: .utility).async { [self] in
-            let boundForVisibleContentKeyNext = PageVisibleContentKey(
-                pageNumber: curPageNum + 1,
-                readingDirection: readingDirection,
-                hMarginDetectStrength: hMarginDetectStrength,
-                vMarginDetectStrength: vMarginDetectStrength
-            )
-            let boundForVisibleContentKeyPrev = PageVisibleContentKey(
-                pageNumber: curPageNum - 1,
-                readingDirection: readingDirection,
-                hMarginDetectStrength: hMarginDetectStrength,
-                vMarginDetectStrength: vMarginDetectStrength
-            )
-            var boundsNext: PageVisibleContentValue? = nil
-            var boundsPrev: PageVisibleContentValue? = nil
-            defer {
-                DispatchQueue.main.async {
-                    if boundsNext != nil {
-                        pageVisibleContentBounds[boundForVisibleContentKeyNext] = boundsNext!
-                    }
-                    if boundsPrev != nil {
-                        pageVisibleContentBounds[boundForVisibleContentKeyPrev] = boundsPrev!
-                    }
-                }
-            }
-            
-            if pageVisibleContentBounds[boundForVisibleContentKeyNext] == nil,
-               let prePage = pdfView.document?.page(at: boundForVisibleContentKeyNext.pageNumber - 1) {   //page(at:) is 0-based
-                boundsNext = getVisibleContentsBound(pdfPage: prePage, readingDirection: readingDirection, hMarginDetectStrength: hMarginDetectStrength, vMarginDetectStrength: vMarginDetectStrength, marginOffset: marginOffset)
-            }
-                        
-            if pageVisibleContentBounds[boundForVisibleContentKeyPrev] == nil,
-               let prePage = pdfView.document?.page(at: boundForVisibleContentKeyPrev.pageNumber - 1) {   //page(at:) is 0-based
-                boundsPrev = getVisibleContentsBound(pdfPage: prePage, readingDirection: readingDirection, hMarginDetectStrength: hMarginDetectStrength, vMarginDetectStrength: vMarginDetectStrength, marginOffset: marginOffset)
-            }
-        }
-        print("\(#function) pageVisibleContentBounds.count=\(pageVisibleContentBounds.count)")
-        
-        if let pageViewPosition = getPageViewPositionHistory(curPageNum),
-           pageViewPosition.scaler > 0,
-           pageViewPosition.viewSize == pdfView.frame.size {
-            let lastDest = PDFDestination(
-                page: curPage,
-                at: pageViewPosition.point
-            )
-            lastDest.zoom = pageViewPosition.scaler
-            print("\(#function) displayMode=\(pdfView.displayMode) BEFORE POINT lastDestPoint=\(lastDest.point)")
-            
-            let bottomRight = PDFDestination(
-                page: curPage,
-                at: CGPoint(x: lastDest.point.x + boundsForCropBox.width, y: lastDest.point.y + boundsForCropBox.height)
-            )
-            bottomRight.zoom = 1.0
-            
-            pdfView.scaleFactor = pageViewPosition.scaler
-            
-            pdfView.go(to: bottomRight)
-            pdfView.go(to: lastDest)
-            return
-        }
-        
-        guard pdfView.scaleFactor > 0 else { return }
-
-        let visibleWidthRatio = 1.0 * (boundForVisibleContent.width + 1) / boundsForCropBox.width
-        let visibleHeightRatio = 1.0 * (boundForVisibleContent.height + 1) / boundsForCropBox.height
-        print("\(#function) curScale scaleFactor=\(pdfView.scaleFactor) visibleWidthRatio=\(visibleWidthRatio) visibleHeightRatio=\(visibleHeightRatio) boundsForCropBox=\(boundsForCropBox) boundForVisibleContent=\(boundForVisibleContent)")
-        
-        let newDestX = boundForVisibleContent.minX + boundsForMediaBox.minX + 2
-        let newDestY = boundsForCropBox.height - boundForVisibleContent.minY + 2
-        
-        let visibleRectInView = pdfView.convert(
-            CGRect(x: newDestX,
-                   y: newDestY,
-                   width: boundsForCropBox.width * visibleWidthRatio,
-                   height: boundsForCropBox.height * visibleHeightRatio),
-            from: curPage)
-        
-        print("\(#function) pdfView pdfView.frame=\(pdfView.frame)")
-        
-        print("\(#function) initialRect visibleRectInView=\(visibleRectInView)")
-        
-        // let insetsScaleFactor = 0.9
-        let insetsHorizontalScaleFactor = 1.0 - (pdfOptions.hMarginAutoScaler * 2.0) / 100.0
-        let insetsVerticalScaleFactor = 1.0 - (pdfOptions.vMarginAutoScaler * 2.0) / 100.0
-        let scaleFactor = { () -> CGFloat in
-            if pdfOptions.lastScale < 0 || pdfOptions.selectedAutoScaler != PDFAutoScaler.Custom {
-                switch pdfOptions.selectedAutoScaler {
-                case .Width:
-                    return pdfView.scaleFactor * pdfView.frame.width / visibleRectInView.width * CGFloat(insetsHorizontalScaleFactor)
-                case .Height:
-                    return pdfView.scaleFactor * pdfView.frame.height / visibleRectInView.height * CGFloat(insetsVerticalScaleFactor)
-                default:    // including .Page
-                    return min(
-                        pdfView.scaleFactor * pdfView.frame.width / visibleRectInView.width * CGFloat(insetsHorizontalScaleFactor),
-                        pdfView.scaleFactor * pdfView.frame.height / visibleRectInView.height * CGFloat(insetsVerticalScaleFactor)
-                    )
-                }
-            } else {
-                return pdfOptions.lastScale
-            }
-        }()
-        pdfView.scaleFactor = scaleFactor
-        //pdfView.scaleFactor = self.lastScale
-//        let navBarFrame = self.navigationController!.navigationBar.frame
-//        let navBarFrameInPDF = pdfView.convert(navBarFrame, to: curPage)
-        let viewFrameInPDF = pdfView.convert(pdfView.frame, to: curPage)
-//        let stackViewFrameInPDF = pdfView.convert(stackView.frame, to: curPage)
-        
-        var newDestPoint = CGPoint(
-            x: newDestX - (1.0 - insetsHorizontalScaleFactor) / 2 * boundsForCropBox.width + boundsForCropBox.minX,
-            y: newDestY /*+ navBarFrameInPDF.height*/ + boundsForCropBox.minY + (1.0 - insetsVerticalScaleFactor) / 2 * viewFrameInPDF.height
-        )
-        
-        if let pageViewPositionHistory = getPageViewPositionHistory(curPageNum) {
-            if pageViewPositionHistory.point.x.isNaN == false {
-                newDestPoint.x = pageViewPositionHistory.point.x
-            }
-            if pageViewPositionHistory.point.y.isNaN == false {
-                newDestPoint.y = pageViewPositionHistory.point.y
-            }
-            print("\(#function) newDest newDestX=\(newDestX) minus=\((1.0 - insetsHorizontalScaleFactor) / 2 * boundsForCropBox.width) plus=\(boundsForCropBox.minX) history=\(pageViewPositionHistory.point.x)")
-            print("\(#function) newDest newDestY=\(newDestX) plus1=\(0) plus2=\(boundsForCropBox.minY) plus3=\((1.0 - insetsVerticalScaleFactor) / 2 * viewFrameInPDF.height) history=\(pageViewPositionHistory.point.y)")
-        } else {
-            print("\(#function) newDest newDestX=\(newDestX) minus=\((1.0 - insetsHorizontalScaleFactor) / 2 * boundsForCropBox.width) plus=\(boundsForCropBox.minX)")
-            print("\(#function) newDest newDestY=\(newDestX) plus1=\(0) plus2=\(boundsForCropBox.minY) plus3=\((1.0 - insetsVerticalScaleFactor) / 2 * viewFrameInPDF.height)")
-        }
-        
-        let newDest = PDFDestination(
-            page: curPage,
-            at: newDestPoint
-        )
-        
-        
-        let initialDestPoint = pdfView.currentDestination!.point
-        
-        print("\(#function) BEFORE POINT curDestPoint=\(pdfView.currentDestination!.point) newDestPoint=\(newDest.point) boundsForCropBox=\(boundsForCropBox)")
-        //            let bottomRight = PDFDestination(
-        //                page: curPage,
-        //                at: CGPoint(x: newDestX + boundsForCropBox.width, y: newDestY + boundsForCropBox.height))
-        
-        let bottomRight = PDFDestination(
-            page: curPage,
-            at: CGPoint(x: boundsForMediaBox.width, y: 0))
-        
-        pdfView.go(to: bottomRight)
-        
-        print("\(#function) BEFORE POINT BOTTOM RIGHT curDestPoint=\(pdfView.currentDestination!.point) newDestPoint=\(newDest.point) boundsForCropBox=\(boundsForCropBox)")
-        
-        pdfView.go(to: newDest)
-        
-        var afterPointX = pdfView.currentDestination!.point.x
-        var afterPointY = pdfView.currentDestination!.point.y /*+ navBarFrameInPDF.height*/ + viewFrameInPDF.height
-        
-        print("\(#function) AFTER POINT scale=\(scaleFactor) curDestPoint=\(pdfView.currentDestination!.point) curDestPointInPDF=\(afterPointX),\(afterPointY) gotoDestPoint=\(newDest.point) boundsForCropBox=\(boundsForCropBox)")
-        
-        let newDestForCompensation = PDFDestination(
-            page: curPage,
-            at: CGPoint(
-                x: newDest.point.x - (afterPointX - newDest.point.x),
-                y: newDest.point.y - (afterPointY - newDest.point.y) - (initialDestPoint.y < 0 ? initialDestPoint.y : 0)
-            )
-        )
-        
-        pdfView.go(to: bottomRight)
-        pdfView.go(to: newDestForCompensation)
-        afterPointX = pdfView.currentDestination!.point.x
-        afterPointY = pdfView.currentDestination!.point.y /*+ navBarFrameInPDF.height*/ + viewFrameInPDF.height
-        print("\(#function) AFTER POINT COMPENSATION scale=\(scaleFactor) curDestPoint=\(pdfView.currentDestination!.point) curDestPointInPDF=\(afterPointX),\(afterPointY) gotoDestPoint=\(newDestForCompensation.point) boundsForCropBox=\(boundsForCropBox)")
-
-        print("\(#function) scaleFactor=\(pdfOptions.lastScale)")
-        
-        #if DEBUG
-        let newDestAnnotation = PDFAnnotation(
-            bounds: .init(origin: newDest.point, size: .init(width: 4, height: 4)),
-            forType: .circle,
-            withProperties: nil
-        )
-        curPage.addAnnotation(newDestAnnotation)
-        
-        let newDestForCompensationAnnotation = PDFAnnotation(
-            bounds: .init(origin: newDestForCompensation.point, size: .init(width: 4, height: 4)),
-            forType: .square,
-            withProperties: nil
-        )
-        newDestForCompensationAnnotation.color = .red
-        curPage.addAnnotation(newDestForCompensationAnnotation)
-        
-        print("\(#function) newDestPoint=\(newDest.point) cropBox=\(curPage.bounds(for: .cropBox)) mediaBox=\(curPage.bounds(for: .mediaBox))")
-        #endif
-        
-        updatePageViewPositionHistory()
-        updateReadingProgress()
-    }
-    
-    func getThumbnailImageSize(boundsForCropBox: CGRect) -> CGSize {
-        // return CGSize(width: boundsForCropBox.width / 4, height: boundsForCropBox.height / 4)
-        if boundsForCropBox.width < 1024 && boundsForCropBox.height < 1024 {
-            return CGSize(width: boundsForCropBox.width, height: boundsForCropBox.height)
-        } else {
-            var width = boundsForCropBox.width
-            var height = boundsForCropBox.height
-            repeat {
-                width /= 2
-                height /= 2
-            } while( width > 1024 || height > 1024)
-            return CGSize(width: width, height: height)
-        }
-    }
-    
-    func getVisibleContentsBound(pdfPage: PDFPage, readingDirection: PDFReadDirection, hMarginDetectStrength: Double, vMarginDetectStrength: Double, marginOffset: Double) -> PageVisibleContentValue {
-        let boundsForMediaBox = pdfPage.bounds(for: .mediaBox)
-        let boundsForCropBox = pdfPage.bounds(for: .cropBox)
-        let sizeForThumbnailImage = getThumbnailImageSize(boundsForCropBox: boundsForCropBox)
-        let thumbnailScale = sizeForThumbnailImage.width / boundsForCropBox.width
-
-        let imageMediaBox = pdfPage.thumbnail(
-            of: CGSize(
-                width: boundsForMediaBox.width * thumbnailScale,
-                height: boundsForMediaBox.height * thumbnailScale),
-            for: .mediaBox
-        )
-        let imageCropBox = pdfPage.thumbnail(of: sizeForThumbnailImage, for: .cropBox)
-        
-        guard let cgimage = imageMediaBox.cgImage else { return PageVisibleContentValue(bounds: boundsForMediaBox, thumbImage: nil) }
-        
-        let numberOfComponents = 4
-        
-        var top = (0, 0)
-        var bottom = (0, 0)
-        var leading = (0, 0)
-        var trailing = (0, 0)
-        
-        print("\(#function) bounds cropBox=\(boundsForCropBox) mediaBox=\(pdfPage.bounds(for: .mediaBox)) artBox=\(pdfPage.bounds(for: .artBox)) bleedBox=\(pdfPage.bounds(for: .bleedBox)) trimBox=\(pdfPage.bounds(for: .trimBox))")
-        print("\(#function) sizeForThumbnailImage \(sizeForThumbnailImage)")
-        print("\(#function) imageCropBox width=\(imageCropBox.size.width) height=\(imageCropBox.size.height)")
-        print("\(#function) imageMediaBox width=\(imageMediaBox.size.width) height=\(imageMediaBox.size.height)")
-
-        let align = 8
-        let padding = (align - Int(imageMediaBox.size.width) % align) % align
-        print("\(#function) CGIMAGE PADDING \(padding)")
-        
-        if let provider = cgimage.dataProvider,
-              let providerData = provider.data,
-              let data = CFDataGetBytePtr(providerData) {
-            switch readingDirection {
-            case .LtR_TtB:
-                top      = getBlankBorderWidth(
-                    size: imageMediaBox.size,
-                    padding: padding,
-                    numberOfComponents: numberOfComponents,
-                    orientation: .up,
-                    data: data,
-                    ratio: boundsForMediaBox.width / boundsForCropBox.width,
-                    hMarginDetectStrength: hMarginDetectStrength
-                )
-                bottom   = getBlankBorderWidth(
-                    size: imageMediaBox.size,
-                    padding: padding,
-                    numberOfComponents: numberOfComponents,
-                    orientation: .down,
-                    data: data,
-                    ratio: boundsForMediaBox.width / boundsForCropBox.width,
-                    hMarginDetectStrength: hMarginDetectStrength
-                )
-                leading  = getBlankBorderWidth(
-                    size: imageMediaBox.size,
-                    padding: padding,
-                    numberOfComponents: numberOfComponents,
-                    orientation: .right,
-                    data: data,
-                    ratio: 3 * imageMediaBox.size.height / Double(Int(imageMediaBox.size.height) - top.1 - bottom.1 + 1),
-                    hMarginDetectStrength: vMarginDetectStrength
-                )
-                trailing = getBlankBorderWidth(
-                    size: imageMediaBox.size,
-                    padding: padding,
-                    numberOfComponents: numberOfComponents,
-                    orientation: .left,
-                    data: data,
-                    ratio: 3 * imageMediaBox.size.height / Double(Int(imageMediaBox.size.height) - top.1 - bottom.1 + 1),
-                    hMarginDetectStrength: vMarginDetectStrength
-                )
-                leading.0 += Int(marginOffset / 100.0 * imageMediaBox.size.width)
-                trailing.0 += Int(marginOffset / 100.0 * imageMediaBox.size.width)
-            case .TtB_RtL:
-                leading  = getBlankBorderWidth(
-                    size: imageMediaBox.size,
-                    padding: padding,
-                    numberOfComponents: numberOfComponents,
-                    orientation: .right,
-                    data: data,
-                    ratio: boundsForMediaBox.height / boundsForCropBox.height,
-                    hMarginDetectStrength: vMarginDetectStrength
-                )
-                trailing = getBlankBorderWidth(
-                    size: imageMediaBox.size,
-                    padding: padding,
-                    numberOfComponents: numberOfComponents,
-                    orientation: .left,
-                    data: data,
-                    ratio: boundsForMediaBox.height / boundsForCropBox.height,
-                    hMarginDetectStrength: vMarginDetectStrength
-                )
-                top      = getBlankBorderWidth(
-                    size: imageMediaBox.size,
-                    padding: padding,
-                    numberOfComponents: numberOfComponents,
-                    orientation: .up,
-                    data: data,
-                    ratio: 3 * imageMediaBox.size.width / Double(Int(imageMediaBox.size.width) - leading.1 - trailing.1 + 1),
-                    hMarginDetectStrength: hMarginDetectStrength
-                )
-                bottom   = getBlankBorderWidth(
-                    size: imageMediaBox.size,
-                    padding: padding,
-                    numberOfComponents: numberOfComponents,
-                    orientation: .down,
-                    data: data,
-                    ratio: 3 * imageMediaBox.size.width / Double(Int(imageMediaBox.size.width) - leading.1 - trailing.1 + 1),
-                    hMarginDetectStrength: hMarginDetectStrength
-                )
-                top.0 += Int(marginOffset / 100.0 * imageMediaBox.size.height)
-                bottom.0 += Int(marginOffset / 100.0 * imageMediaBox.size.height)
-                break
-            }
-            
-        }
-        
-        print("\(#function) white border page=\(pdfPage.pageRef!.pageNumber) \(top) \(bottom) \(leading) \(trailing)")
-        
-        UIGraphicsBeginImageContextWithOptions(imageMediaBox.size, false, CGFloat.zero)
-        imageMediaBox.draw(at: CGPoint.zero)
-        
-        //imageCropBox.draw(at: CGPoint(x: boundsForCropBox.minX, y: boundsForMediaBox.maxY - boundsForCropBox.maxY), blendMode: .darken, alpha: 0.8)
-        
-        let rectangle = CGRect(
-            x: leading.0,
-            y: top.0,
-            width: trailing.0 - leading.0 + 2,
-            height: bottom.0 - top.0 + 1
-        )
-        UIColor.black.setFill()
-        UIRectFrame(rectangle)
-        
-        #if DEBUG
-        UIColor.red.setStroke()
-        let drawBounds = CGRect(x: boundsForCropBox.minX * thumbnailScale,
-                                y: boundsForCropBox.minY * thumbnailScale,
-                                width: sizeForThumbnailImage.width,
-                                height: sizeForThumbnailImage.height)
-        UIRectFrame(drawBounds)
-        #endif
-        
-        let newImage = UIGraphicsGetImageFromCurrentImageContext()
-        UIGraphicsEndImageContext()
-        
-//        return CGRect(x: rectangle.minX / thumbnailScale, y: rectangle.minY / thumbnailScale, width: rectangle.width / thumbnailScale, height: rectangle.height / thumbnailScale)
-        //transform to cropBox coordination
-        return PageVisibleContentValue(
-            bounds: CGRect(
-                x: CGFloat(leading.0) / thumbnailScale - boundsForCropBox.minX,
-                y: CGFloat(top.0) / thumbnailScale - (boundsForMediaBox.maxY - boundsForCropBox.maxY),
-                width: rectangle.width / thumbnailScale,
-                height: rectangle.height / thumbnailScale
-            ),
-            thumbImage: newImage
-        )
-    }
-    
-    /*
-     from top to bottom
-     */
-    func getBlankBorderWidth(size: CGSize, padding: Int, numberOfComponents: Int, orientation: CGImagePropertyOrientation, data: UnsafePointer<UInt8>, ratio: Double = 1.0, hMarginDetectStrength: Double) -> (Int, Int) {
-        let lineNumMax = { () -> Int in
-            switch(orientation) {
-            case .up, .down, .upMirrored, .downMirrored:
-                return Int(size.height)
-            case .left, .leftMirrored, .right, .rightMirrored:
-                return Int(size.width)
-            }
-        }()
-        let pixelNumMax = { () -> Int in
-            switch(orientation) {
-            case .up, .down, .upMirrored, .downMirrored:
-                return Int(size.width)
-            case .left, .leftMirrored, .right, .rightMirrored:
-                return Int(size.height)
-            }
-        }()
-        var border = lineNumMax/2
-        let pixelNumInRow = Int(size.width) + padding
-        var nonWhiteLineFirst = 0
-        var nonWhiteLines = 0
-        var whiteLines = 0
-        for line in (1 ..< (lineNumMax/4)) {    //bypassing first line due to noises
-            var nonWhiteDensity = 0.0
-            for pixelInLine in (1 ..< pixelNumMax) {    //bypassing first line due to noises
-                let lineIndex = { () -> Int in
-                    switch(orientation) {
-                    case .up, .upMirrored, .right, .rightMirrored:
-                        return line     //top to bottom & left to right
-                    case .down, .downMirrored, .left, .leftMirrored:
-                        return lineNumMax - line - 1
-                    }
-                }()
-                
-                let pixelIndex = { () -> Int in
-                    switch(orientation) {
-                    case .up, .down, .upMirrored, .downMirrored:
-                        return pixelInLine + pixelNumInRow * lineIndex
-                    case .left, .leftMirrored, .right, .rightMirrored:
-                        return lineIndex + pixelNumInRow * pixelInLine  //pixelInLine is row number of data
-                    }
-                }() * numberOfComponents
-                
-                
-//                let pixelIndex = (
-//                    ( Int(size.width) + padding ) * lineIndex   //locating target line in data
-//                        + pixelInLine                           //moving to pixel
-//                ) * numberOfComponents
-                
-                nonWhiteDensity += getPixelGreyLevel(pixelIndex: pixelIndex, data: data)
-            }
-            if nonWhiteDensity > 0 {
-                print("nonWhiteDensity h=\(line) density=\(nonWhiteDensity) orientation=\(orientation.rawValue)")
-            }
-            
-            if nonWhiteDensity > 0,
-               nonWhiteDensity / Double(pixelNumMax) * ratio * 20.0 > hMarginDetectStrength {
-                nonWhiteLines += 1
-                if nonWhiteLineFirst == 0 {
-                    nonWhiteLineFirst = line
-                }
-            } else {
-                whiteLines += 1
-                nonWhiteLines = 0
-                nonWhiteLineFirst = 0
-            }
-            
-                //print("isWhite h=\(h) \(isWhite)")
-            if nonWhiteLines > 2,
-               nonWhiteLineFirst < lineNumMax / 4,
-               border == lineNumMax / 2 {
-                border = nonWhiteLineFirst
-            }
-        }
-        
-        if border == lineNumMax/2 {
-            border = 1
-        }
-        
-        switch(orientation) {
-        case .up, .upMirrored, .right, .rightMirrored:
-            return (border, whiteLines)     //top to bottom & left to right
-        case .down, .downMirrored, .left, .leftMirrored:
-            return (lineNumMax - border - 1, whiteLines)
-        }
-        
-//        if bottomUp {
-//            border = Int(size.height) - border - 1
-//        }
-        
-//        return border
-        
-    }
-    
-    func getPixelGreyLevel(pixelIndex: Int, data: UnsafePointer<UInt8>) -> Double {
-        let r = data[pixelIndex]
-        let g = data[pixelIndex + 1]
-        let b = data[pixelIndex + 2]
-        
-        
-        if r < 200 && g < 200 && b < 200 {
-            return Double(UInt(255-r) + UInt(255-g) + UInt(255-b)) / 3 / 255.0
-            //print("top=\(h) w=\(w) \(r) \(g) \(b) \(a)")
-            //isWhite = false
-            //break
-        } else {
-            return 0.0
-        }
-    }
-    
     func addBlankSubView(page: PDFPage?) {
-        let backgroundColor = UIColor(cgColor: PDFPageWithBackground.fillColor ?? CGColor.init(gray: 1.0, alpha: 1.0))
-        
-        if blankView.layer.compositingFilter == nil,
-           let page = page as? PDFPageWithBackground {
-            let bounds = pageVisibleContentBounds[
-                PageVisibleContentKey(
-                    pageNumber: page.pageRef?.pageNumber ?? -1,
-                    readingDirection: pdfOptions.readingDirection,
-                    hMarginDetectStrength: pdfOptions.hMarginDetectStrength,
-                    vMarginDetectStrength: pdfOptions.vMarginDetectStrength
-                )
-            ]?.bounds ?? .zero
-            
-            let thumbImage = page.thumbnailWithBackground(of: pdfView.frame.size, for: .cropBox, by: bounds)
-            blankView.image = thumbImage
-        }
-        
-        blankView.tintColor = backgroundColor
-        blankView.backgroundColor = backgroundColor
-        blankView.frame.size = pdfView.frame.size
-        
-        blankActivityView.frame = CGRect(x: pdfView.frame.width/2, y: pdfView.frame.height/2, width: 50, height: 50)
-        blankActivityView.style = .large
-        blankActivityView.backgroundColor = .clear
-        blankActivityView.startAnimating()
-        
-        clearBlankSubView()
+        marginCropController.showBlankOverlay(page: page, options: pdfOptions)
     }
     
     func clearBlankSubView() {
-        DispatchQueue.main.asyncAfter(deadline: .now().advanced(by: .milliseconds(400))) {
-            self.blankView.frame.size = .zero
-            self.blankActivityView.stopAnimating()
-        }
-    }
-    
-    func handleOptionsChange(pdfOptions: PDFOptions) {
-        print(pdfOptions)
-        
-        let oldOptions = PDFOptions(value: self.pdfOptions)
-        
-        if self.pdfOptions !== pdfOptions {
-            if let realm = self.pdfOptions.realm {
-                try? realm.write {
-                    self.pdfOptions.update(other: pdfOptions)
-                }
-            } else {
-                self.pdfOptions = pdfOptions
-                return
-            }
-        }
-        
-        if oldOptions.pageMode != self.pdfOptions.pageMode || oldOptions.scrollDirection != self.pdfOptions.scrollDirection {
-            updatePageViewPositionHistory()
-        }
-        
-        // Trigger didSet for UI refresh (colors, etc.)
-        self.pdfOptions = self.pdfOptions
-        
-        if oldOptions.themeMode != self.pdfOptions.themeMode {
-//                self.pdfView.layoutDocumentView()
-            //self.pdfView.invalidateIntrinsicContentSize()
-            let scaleFactor = self.pdfView.scaleFactor
-            self.pdfView.scaleFactor = 1.0
-            self.pdfView.scaleFactor = scaleFactor
-        }
-        if let pageNum = pdfView.currentPage?.pageRef?.pageNumber {
-            self.pageViewPositionHistory[pageNum]?.scaler = 0
-            switch self.pdfOptions.readingDirection {
-            case .LtR_TtB:
-                self.pageViewPositionHistory[pageNum]?.point.x = .nan
-            case .TtB_RtL:
-                self.pageViewPositionHistory[pageNum]?.point.y = .nan
-            }
-        }
-        handlePageChange(notification: Notification(name: .PDFViewScaleChanged))
-    }
-    
-    func handleAutoScalerChange(autoScaler: PDFAutoScaler, hMarginAutoScaler: Double, vMarginAutoScaler: Double) {
-        
-    }
-    
-    @objc func handleScaleChange(_ sender: Any?)
-    {
-        let newScale = pdfView.scaleFactor
-        guard abs(pdfOptions.lastScale - newScale) > 0.0001 else { return }
-        
-        if let realm = pdfOptions.realm {
-            try? realm.write {
-                pdfOptions.lastScale = newScale
-            }
-        } else {
-            pdfOptions.lastScale = newScale
-        }
-        print("handleScaleChange: \(pdfOptions.lastScale)")
-    }
-    
-    @objc func handleDisplayBoxChange(_ sender: Any?)
-    {
-        print("handleDisplayBoxChange: \(self.pdfView.currentDestination!)")
-    }
-    
-    @objc func finishReading(sender: UIBarButtonItem) {
-        updatePageViewPositionHistory()
-
-        updateReadingProgress()
-        
-        
-        self.dismiss(animated: true) {
-            self.pdfView.document = nil
-            self.yabrPDFMetaSource = nil
-            self.tocList.removeAll()
-        }
-    }
-    
-    func updateReadingProgress() {
-        var position = [String : Any]()
-        
-        guard let curPageNum = pdfView.page(for: .zero, nearest: true)?.pageRef?.pageNumber,
-              let curPagePos = getPageViewPositionHistory(curPageNum)
-        else { return }
-        
-        position["pageNumber"] = curPageNum
-        position["pageOffsetX"] = curPagePos.point.x
-        position["pageOffsetY"] = curPagePos.point.y
-        
-        let bookProgress = 100.0 * Double(position["pageNumber"] as? Int ?? 0) / Double(pdfView.document?.pageCount ?? 1)
-        
-        var chapterProgress = 0.0
-        let chapterName = titleInfoButton.currentTitle ?? "Unknown Title"
-        if let firstIndex = tocList.lastIndex(where: { $0.0 == chapterName && $0.1 <= curPageNum }) {
-            let curIndex = firstIndex.advanced(by: 0)
-            let nextIndex = firstIndex.advanced(by: 1)
-            let chapterStartPageNum = tocList[curIndex].1
-            let chapterEndPageNum = nextIndex < tocList.count ?
-                tocList[nextIndex].1 + 1 : (pdfView.document?.pageCount ?? 1) + 1
-            if chapterEndPageNum > chapterStartPageNum {
-                chapterProgress = 100.0 * Double(curPageNum - chapterStartPageNum) / Double(chapterEndPageNum - chapterStartPageNum)
-            }
-        }
-        
-        let enginePos = ReaderEnginePosition(
-            pageNumber: curPageNum,
-            maxPage: self.pdfView.document?.pageCount ?? 1,
-            pageOffsetX: Int(curPagePos.point.x.rounded()),
-            pageOffsetY: Int(curPagePos.point.y.rounded()),
-            bookProgress: bookProgress,
-            chapterProgress: chapterProgress,
-            chapterName: chapterName,
-            cfi: nil
-        )
-        self.readerEngineDelegate?.readerEngine(self, didUpdatePosition: enginePos)
-    }
-    
-//    @objc func lookupStarDict() {
-//        if let s = pdfView.currentSelection?.string {
-//            print(s)
-//            starDictView.word = s
-//            self.present(starDictView, animated: true, completion: nil)
-//        }
-//    }
-    
-    func updatePageViewPositionHistory() {
-        guard let pagePoint = getPagePoint() else { return }
-        
-        pageViewPositionHistory[pagePoint.0] = pagePoint.1
-        print("updatePageViewPositionHistory \(pagePoint)")
-    }
-    
-    func getPagePoint() -> (Int, PageViewPosition)? {
-        guard let curDest = pdfView.currentDestination,
-              let curDestPage = curDest.page,
-              let curPage = pdfView.page(for: .zero, nearest: true),
-              let curPageNum = curPage.pageRef?.pageNumber
-              else { return nil }
-        
-        var curDestPoint = curDest.point
-        if curPage != curDestPage {
-//            CGPoint(x: curDest.point.x, y: curDest.point.y - curDestPage.bounds(for: .cropBox).height)
-            //translate curDest point to curPage point
-            let curDestPointInView = pdfView.convert(curDestPoint, from: curDestPage)
-            let curDestPointInCurPage = pdfView.convert(curDestPointInView, to: curPage)
-            curDestPoint = curDestPointInCurPage
-        }
-        
-        let viewFrameInPDF = pdfView.convert(pdfView.frame, to: curPage)
-//        let navBarFrame = self.navigationController?.navigationBar.frame ?? CGRect()
-//        let navBarFrameInPDF = pdfView.convert(navBarFrame, to:curPage)
-    
-        let pointUpperLeft = CGPoint(
-            x: curDestPoint.x,
-            y: curDestPoint.y /*+ navBarFrameInPDF.height*/ + viewFrameInPDF.height
-        )
-        
-        return (
-            curPageNum,
-            PageViewPosition(
-                scaler: pdfView.scaleFactor,
-                point: pointUpperLeft,
-                viewSize: pdfView.frame.size
-            )
-        )
-    }
-    
-    func getPageViewPositionHistory(_ pageNum: Int) -> PageViewPosition? {
-        return self.pageViewPositionHistory[pageNum]
-    }
-    
-    func buildDefaultMenuItems() -> [UIMenuItem] {
-        let highlightMenuItem = UIMenuItem(title: "HighlightA", action: #selector(highlightAction(_:)))
-        
-        var menuItems = [highlightMenuItem]
-        
-        if let dictViewer = yabrPDFMetaSource?.yabrPDFDictViewer(pdfView) {
-            let dictViewerItem = UIMenuItem(title: "MDict", image: UIImage(systemName: "character.book.closed")) { [weak self] _ in
-                self?.dictViewerAction(self)
-            }
-//            menuItems.append(UIMenuItem(title: dictViewer.0, action: #selector(dictViewerAction)))
-            menuItems.append(dictViewerItem)
-            dictViewer.1.loadViewIfNeeded()
-        }
-        
-        return menuItems
-    }
-    
-    @objc func highlightAction(_ sender: Any?) {
-        if let highlightTapped = pdfView.highlightTapped {
-            
-        } else {
-            guard let currentSelection = pdfView.currentSelection else { return }
-            
-            var style = BookHighlightStyle.yellow.rawValue
-            if (sender as? UIButton) == annotationView.underlineButton {
-                style = BookHighlightStyle.underline.rawValue
-            }
-            
-            self.annotationManager.addHighlight(style: style, selection: currentSelection)
-            self.annotationView.isHidden = true
-        }
-    }
-    
-    @objc func dictViewerAction(_ sender: Any?) {
-        guard let selectedText = pdfView.currentSelection?.string,
-              let (name, dictViewer) = yabrPDFMetaSource?.yabrPDFDictViewer(pdfView) else { return }
-        
-        print("\(#function) word=\(selectedText)")
-        dictViewer.title = selectedText
-        
-        present(dictViewer, animated: true)
-    }
-    
-    func sharePDF(annotated: Bool) {
-        let provider = YabrPDFSharingProvider(placeholderItem: "")
-        guard let pdfURL = yabrPDFMetaSource?.yabrPDFURL(pdfView),
-              let bookInShelfId = yabrPDFMetaSource?.yabrPDFBook(pdfView, info: "Key")
-        else {
-            return
-        }
-        
-        let bookTitle = yabrPDFMetaSource?.yabrPDFBook(pdfView, info: "Title") ?? "No Title"
-        let bookAuthor = yabrPDFMetaSource?.yabrPDFBook(pdfView, info: "Author") ?? "Unknown"
-        
-        let tmpDir = FileManager.default
-            .temporaryDirectory
-            .appendingPathComponent(bookInShelfId, isDirectory: true)
-        let tmpFile = tmpDir.appendingPathComponent("\(bookTitle) - \(bookAuthor).pdf", isDirectory: false)
-        
-        do {
-            if false == FileManager.default.fileExists(atPath: tmpDir.path) {
-                try FileManager.default.createDirectory(at: tmpDir, withIntermediateDirectories: false)
-            }
-            if FileManager.default.fileExists(atPath: tmpFile.path) {
-                try FileManager.default.removeItem(at: tmpFile)
-            }
-            if annotated {
-                let fillColor = PDFPageWithBackground.fillColor
-                PDFPageWithBackground.fillColor = nil
-                defer {
-                    PDFPageWithBackground.fillColor = fillColor
-                }
-                guard pdfView.document?.write(to: tmpFile) == true
-                else {
-                    return
-                }
-            } else {
-                try FileManager.default.linkItem(at: pdfURL, to: tmpFile)
-            }
-        } catch {
-            print("Save Original PDF error=\(error)")
-            let alert = UIAlertController(title: "Error Sharing PDF", message: error.localizedDescription, preferredStyle: .alert)
-            alert.addAction(.init(title: "OK", style: .default))
-            self.present(alert, animated: true)
-            return
-        }
-        
-        provider.fileURL = tmpFile
-        provider.subject = "Save Original PDF"
-        
-        let vc = UIActivityViewController(activityItems: [provider], applicationActivities: nil)
-        if let popover = vc.popoverPresentationController {
-            popover.barButtonItem = shareBarButtonItem
-        }
-        
-        present(vc, animated: true, completion: nil)
+        marginCropController.hideBlankOverlay()
     }
     
 }
@@ -1662,56 +308,4 @@ protocol YabrPDFMetaSource {
     func yabrPDFReferenceText(_ view: YabrPDFView?, set refText: String?)
     
     func yabrPDFOptionsIsNight<T>(_ view: YabrPDFView?, _ f: T, _ l: T) -> T
-}
-
-@available(iOS 16.0, *)
-extension YabrPDFViewController: UIEditMenuInteractionDelegate {
-    @objc func didLongPress(_ recognizer: UIGestureRecognizer) {
-        let location = recognizer.location(in: pdfView)
-        
-        let configuration = UIEditMenuConfiguration(identifier: nil, sourcePoint: location)
-
-        guard let interaction = pdfView.interactions.first(where: { (($0 as? UIEditMenuInteraction)?.delegate as? YabrPDFViewController) == self }) as? UIEditMenuInteraction
-        else {
-            return
-        }
-        
-//        let aoi = pdfView.areaOfInterest(for: location)
-//        guard aoi.contains(.textArea)
-//        else { return }
-        
-        pdfView.visiblePages.forEach { page in
-            let pagePoint = pdfView.convert(location, to: page)
-            if let pageSelection = pdfView.currentPage?.selectionForWord(at: pagePoint) {
-                pdfView.setCurrentSelection(pageSelection, animate: true)
-                print("\(#function) selection=\(pageSelection)")
-            }
-        }
-        
-        // Present the edit menu interaction.
-        interaction.presentEditMenu(with: configuration)
-    }
-}
-
-extension YabrPDFViewController: ReaderEngineController {
-    func applyPreferences(_ preferences: ReaderEnginePreferences) {
-        let newOptions = PDFOptions()
-        switch preferences.themeMode {
-        case 1:
-            newOptions.themeMode = .serpia
-        case 2:
-            newOptions.themeMode = .dark
-        default:
-            newOptions.themeMode = .none
-        }
-        
-        newOptions.pageMode = preferences.scroll ? .Scroll : .Page
-        newOptions.scrollDirection = preferences.scrollDirection == 0 ? .Vertical : .Horizontal
-        
-        self.handleOptionsChange(pdfOptions: newOptions)
-    }
-
-    func applyHighlights(_ highlights: [ReaderEngineHighlight]) {
-        self.annotationManager.applyHighlights(highlights)
-    }
 }
